+++ R/strat.R
@@ -10,14 +10,17 @@
 #' Then this iteration with the updated group of age estimates is either accepted or rejected. The acceptance probability depends on the iteration's energy; if its energy is higher than that of the previous iteration it is accepted, but if it is lower, it is accepted with a probability proportional to its relative energy. Therefore, over many iterations the process will 'learn' from the data and find high-energy combinations of parameter values that fit with the prior constraints that the ages should be ordered chronologically.
 #' Because the iterations are based on a process of modifying values of one parameter each iteration, and because some iterations will not be accepted, the MCMC output will often have a large degree of dependence between neighbouring iterations. Therefore, some thinning will have to be done, by storing only one every few iterations (default 20). Also, since the initial ball-park estimates could be quite wrong, the first 100 or so iterations should also be discarded (burnin). 
 #' It is thus important to check the time-series of the energy after the run. We don't want to see a remaining burn-in at the start, and we don't want to see a noticeable 'structure' where iterations remain in approximately or entirely the same spot for a long time. Instead, an ideal run will look like white noise.
+#' By default, the model output and the settings are stored in a list called `info'. For example, to retrieve the model output, type info$output. This has a column for each date, and a row for each stored MCMC iteration. The MCMC's energy can be found in info$Us. The model's `structure' such as blocks or gaps can be found in info$struc. To check which dates were used, type info$dets or info$dat (the latter will include all information, including any gaps). 
 #' @param name Name of the stratigraphy dataset. Defaults to \code{"mystrat"}.
-#' @param strat.dir The directory where the folders of the individual stratigraphies live. Defaults to \code{treedir="strats"}.
+#' @param strat.dir The directory where the folders of the individual stratigraphies live. Defaults to \code{strat.dir="strats"}.
 #' @param its Amount of iterations to be run. Setting this to low numbers (e.g., 1000) will result in fast but less stable and less reliable runs. Higher values will take longer but result in more stable and robust runs. Defaults to \code{50000}. Aim to set this to such values that at least 3000 iterations remain after removing the burnin and thinning.
 #' @param burnin Amount of iterations to remove at the start of the run. Defaults to \code{100}.
 #' @param thinning After running all iterations, only some will be stored. For example, if thinning is set at the default \code{50}, only every 50th MCMC iteration will be stored, and the others will be discarded. This is to remove the dependence between neighbouring MCMC iterations. Defaults to a value calculated from the MCMC run itself.
 #' @param internal.thinning Does internal thinning during the MCMC process, storing only every 'internal.thinning' MCMC iterations.
-#' @param min.its Minimum number of (remaining) iterations, below which a warning is given. Defaults for now to 3,000 iterations.
-#' @param write.MCMC Especially longer runs or sites with many dates can take up lots of memory. For such cases, MCMC iterations are stored in temporary files rather than in memory. Defaults to TRUE. 
+#' @param min.its Minimum number of (remaining) iterations, below which a warning is given. Defaults for now to 1,000 iterations.
+#' @param write.MCMC Especially longer runs or sites with many dates can take up lots of memory. For such cases, MCMC iterations are stored in temporary files (and in the session's tempdir() folder) rather than in memory - however this could impact your computer as the files have to be written to repeatedly and they can become huge. Defaults to FALSE. 
+#' @param MCMC.dir Directory where temporary MCMC files will be saved. If left empty, defaults to \code{tempdir()}.
+#' @param remove.tmp If \code{write.MCMC=TRUE}, then by default the (often huge) temporary files are deleted after the run. Set to \code{remove.tmp=FALSE} to keep these temporary files. The files will be placed in the temporary folder of this R session - where this is can be checked by providing the command \code{tempdir()}. 
 #' @param init.ages By default, the ballpark age estimates to feed the MCMC are calculated automatically, however they can also be provided manually, as two rows of values (for x0 and xp0) which have to obey the assumptions (e.g., chronological ordering).
 #' @param ballpark.method Initial, ballpark values for the initial ages (if not provided by the option `init.ages`). Can be 1 (based on a linear model) or 2 (based on sorted random draws).
 #' @param y.scale The scale of the vertical axis of the main plot. This can be the positions of the dated levels (`positions`) or their position order (`dates`). 
@@ -28,17 +31,17 @@
 #' @param delta.STD The error of the offset. Set to 0 by default.
 #' @param t.a First parameter for the student-t distribution (defaults to 3; higher numbers make the distribution approximate the normal distribution more).
 #' @param t.b Second parameter for the student-t distribution (defaults to 4; higher numbers make the distribution 
-#' @param cc Calibration curve to be used. Could be 1 (IntCal20; default), 2 (Marine20), 3 (SHCal20) or 4 (custom curve).
-#' @param cc.dir Directory of calibration curve. Keep empty for the default value.
+#' @param cc.dir Directory of calibration curve(s). Keep empty for the default value.
+# @param prob After the run, a fit of the model with the dates is calculated, as the ratio of model iterations that fit the hpd ranges of the dates. Defaults to the \code{prob=0.95} hpd ranges.
 #' @param postbomb Negative C-14 ages should be calibrated using a postbomb curve. This could be 1 (northern-hemisphere region 1), 2 (NH region 2), 3 (NH region 3), 4 (southern hemisphere regions 1-2), or 5 (SH region 3).
 #' @param BCAD The calendar scale of graphs and age output-files is in \code{cal BP} by default, but can be changed to BC/AD using \code{BCAD=TRUE}. Needs more work probably.
 #' @param ask Whether or not to ask if a folder should be made (if required).
 #' @param talk Whether or not to provide feedback on folders written into and on what is happening.
-#' @param talk Whether or not show the progress of the MCMC run.
 #' @param clean.garbage Whether or not to clean up the memory 'garbage collection' after a run. Recommendable if you have many dates or long runs.
+#' @param save.info Whether or not to store a variable `info' in the session which contains the run input, output and settings. Defaults to \code{save.info=TRUE}.
 #' @param age.span Expected age span. Defaults to run from the current year in AD to 55e3 which is the current cal BP limit for C-14 dates. If older, non-14C dates are present, age.span is set to the larger of the radiocarbon limit or twice the age of the oldest non-radiocarbon age.
-#' @param ... Options for the plot. See \code{plot.strat}.
-#' @return a variable 'info' which contains the dating and modelling information to produce a plot. Also calls the function \code{draw.strat} to produce a plot of the results.
+#' @param ... Options for the plot. See \code{draw.strat}.
+#' @return a variable 'info' which contains the dating and modelling information to produce a plot (see details). Also calls the function \code{draw.strat} to produce a plot of the results.
 #' @examples
 #' \dontrun{
 #' tmp <- tempdir()
@@ -55,7 +58,7 @@
 #'
 #' Nicholls G, Jones M 2001. Radiocarbon dating with temporal order constraints. Journal of the Royal Statistical Society: Series C (Applied Statistics) 50, 503-521.
 #' @export
-strat <- function(name="mystrat", strat.dir="strats", its=5e4, burnin=100, thinning=c(), internal.thinning=c(), min.its=3e3, write.MCMC=TRUE, init.ages=c(), ballpark.method=2, y.scale="dates", showrun=FALSE, sep=",", normal=FALSE, delta.R=0, delta.STD=0, t.a=3, t.b=4, cc=1, cc.dir=c(), postbomb=FALSE, BCAD=FALSE, ask=FALSE, talk=TRUE, show.progress=TRUE, clean.garbage=TRUE, age.span=c(), ...) {
+strat <- function(name="mystrat", strat.dir="strats", run=TRUE, its=5e4, burnin=100, thinning=c(), internal.thinning=c(), min.its=1e3, write.MCMC=FALSE, MCMC.dir=c(), remove.tmp=TRUE, init.ages=c(), ballpark.method=2, y.scale="dates", showrun=FALSE, sep=",", normal=FALSE, delta.R=0, delta.STD=0, t.a=3, t.b=4, cc.dir=c(), prob=0.95, postbomb=FALSE, BCAD=FALSE, ask=FALSE, talk=TRUE, show.progress=TRUE, clean.garbage=TRUE, save.info=TRUE, age.span=c(), ...) {
   start.time <- as.numeric(format(Sys.time(), "%s"))
   info <- read.strat(name, strat.dir, sep, normal, delta.R, delta.STD, t.a, t.b, cc)
   dat <- info$dets
@@ -79,7 +82,8 @@ strat <- function(name="mystrat", strat.dir="strats", its=5e4, burnin=100, thinn
 
   # set reasonable maximum age based on the dates
   if(length(age.span) == 0) {
-    age.span <- max(2*(dets[,2]+dets[,3]), 55e3) # C14 or cal BP ages
+    dates <- which(dets[,5] %in% 0:4)
+    age.span <- max(2*(dets[dates,2]+dets[dates,3]), 55e3) # C14 or cal BP ages
     age.span <- age.span+(as.numeric(format(Sys.time(), "%Y"))-1950)
   }
 
@@ -128,8 +132,6 @@ strat <- function(name="mystrat", strat.dir="strats", its=5e4, burnin=100, thinn
     init.ages <- ballpark.strat(method=ballpark.method, dat=dets, gaps=struc$gaps, postbomb=postbomb, normal=normal, t.a=t.a, t.b=t.b, cc.dir=cc.dir)
   x0 <- init.ages[1,]; xp0 <- init.ages[2,]
 
-  dets <<- dets; x0 <<- x0; info <<- info; struc <<- struc # tmp
-
   # calculates how well the proposed ages x fit with the dates and any constraints on ordering and gaps:
   energy <- function(x, Dets=dets, Normal=normal, ta=t.a, tb=t.b, M=struc$m,
     curves=ccs, cc1=cc.1, cc2=cc.2, cc3=cc.3, cc4=cc.4, Age.span=age.span,
@@ -150,24 +152,46 @@ strat <- function(name="mystrat", strat.dir="strats", its=5e4, burnin=100, thinn
     x.cc3 <- cc.energy(3, Dets, Dets.cc3, cc3, x, curves, Normal, ta, tb) # SH
     x.cc4 <- cc.energy(4, Dets, Dets.cc4, cc4, x, curves, Normal, ta, tb) # custom ccurve
 
-    if(has.blocks) { # then check for reversals, and adapt from and to to calculate correct l.span
-	  reversal <- FALSE  
+    ### bugs:
+    ### plotted positions of any blocks beyond the first one (incorrectly assigning 'b' in structure function)
+    ### exact gap if it goes counter to dates causes problems (check initvals)
+    ### cal BP date (cc=0)
+
+    ### feature:
+    ### add exact date as option. cc=0, error=0. How separate from other cc=0 dates?
+    ### see if some of the variables in the options of energy can be removed (faster)?
+    ### redo overlap by calculating the product of calibrated*modelled, over the same yrs? And then?
+
+    if(has.blocks) { # then check for reversals, and adapt 'from' and 'to' to calculate correct l.span
+      reversal <- 0
+
       for(i in 1:length(blocks)) {
         thisblock <- within.block[[i]]
-        if (x[above.block] > min(x[thisblock]) || x[below.block] < max(x[thisblock]) ) {
-          reversal <- TRUE
-		  break
-		} else {
-		  o <- order(x[thisblock])
-          to[above.block] <- thisblock[o[1]] # position of youngest age 
-          from[below.block] <- thisblock[o[length(o)]] # ... and of oldest age
-        } 
-	  }
-	  if(reversal)
+        if((x[above.block[i]] > min(x[thisblock])) ) {
+          # cat("reversal on top!\n")
+          reversal <- reversal + 1
+          break
+        }
+
+        if(x[below.block[i]] < max(x[thisblock])) {
+          # cat("reversal at bottom!\n")
+          reversal <- reversal + 1
+          break
+        }
+        o <- order(x[thisblock])
+        to[above.block[[i]]] <- thisblock[o[1]]
+        from[below.block[[i]]] <- thisblock[o[length(o)]] # ... and of oldest age
+      }
+#	  if(reversal > 0)
+#		  cat("!") else cat(".")
+      if(reversal > 0)
         l.x <- -Inf else
           l.x <- mapply(l.span, x[from], x[to], Age.span, M)
+
     } else
       l.x <- mapply(l.span, x[from], x[to], Age.span, M)
+
+    # l.x <<- l.x; struc <<- struc; info <<- info; x.cc1 <<- x.cc1; x.cc0 <<- x.cc0; x <<- x; Age.span <<- Age.span; M <<- M
 
     # now recalculate the energy between dates with any provided gap information
     if(has.gaps) {
@@ -191,43 +215,76 @@ strat <- function(name="mystrat", strat.dir="strats", its=5e4, burnin=100, thinn
   support <- function(ages)
     return(!min(diff(ages)) < 0)
 
-  # write output to files if required
+  # if there are blocks, order the ages within them (even though they don't have to be)
+  if(struc$has.blocks)
+    support <- function(ages) {
+      inblock <- struc$within.block
+      for(i in 1:length(inblock))
+        ages[inblock[[i]]] <- sort(ages[inblock[[i]]])
+      return(!min(diff(ages)) < 0)
+    }
+
+  # write temporary output to files if required (these can become huge!)
   out.fl <- c(); energy.fl <- c()
   if(write.MCMC) {
-    out.fl <- file.path(strat.dir, name, paste0(name, "_tmp.out"))
-    energy.fl <- file.path(strat.dir, name, paste0(name, "_tmp_energy.out"))
+	if(length(MCMC.dir) == 0)
+	  MCMC.dir <- tempdir()	  
+    out.fl <- file.path(MCMC.dir, paste0(name, "_tmp.out"))
+    energy.fl <- file.path(MCMC.dir, paste0(name, "_tmp_energy.out"))
+    if(file.exists(out.fl))
+      file.remove(out.fl)
+    if(file.exists(energy.fl))
+      file.remove(energy.fl)
   }
 
-  # run the twalk, but storing only output and energy (Us)
-  tw <- Runtwalk(Tr=its, Obj=energy, Supp=support, dim=nrow(dets), x0=x0, xp0=xp0,
-    thinning=internal.thinning, out.fl=out.fl, energy.fl=energy.fl, show.progress=show.progress)
-  output <- tw$output; Us <- tw$Us;
+  if(!run) { # then we read the .out file (if it exists)
+    out.fl <- file.path(strat.dir, name, paste0(name, ".out"))
+    energy.fl <- file.path(strat.dir, name, paste0(name, "_energy.out"))
+    if(!file.exists(out.fl) || !file.exists(energy.fl))
+      stop("Cannot find the .out and _energy.out files of site", name)
+    output <- fastread(out.fl, header=FALSE)
+    Us <- fastread(energy.fl, header=FALSE)[,1]
+  } else {
 
-  # post-run, remove any burn-in iterations
-  if(burnin > 0) {
-    output <- output[-(1:burnin),]
-    Us <- Us[-(1:burnin)]
+    # run the twalk, storing only output and energy (Us)
+    tw <- Runtwalk(Tr=its, Obj=energy, Supp=support, dim=nrow(dets), x0=x0, xp0=xp0,
+      thinning=internal.thinning, out.fl=out.fl, energy.fl=energy.fl, show.progress=show.progress)
+    output <- tw$output; Us <- tw$Us;
+
+    if(write.MCMC)
+      if(remove.tmp) { # these temporary files can become huge!
+        file.remove(out.fl) 
+        file.remove(energy.fl)
+      }
+    # post-run, remove any burn-in iterations
+    if(burnin > 0) {
+      output <- output[-(1:burnin),]
+      Us <- Us[-(1:burnin)]
+    }
+    message("Removed a burn-in of ", burnin)
+
+    # additional thinning after the run
+	# by default, takes the thinning value as suggested by rtwalk's IAT,
+	# but divided by 5 because proposed IAT values often result in very much thinning 
+    if(length(thinning) == 0) 
+      thinning <- ceiling(coffee::IAT(tw, 0, burnin, nrow(output))/5)
+    message("Thinning the MCMC by storing every ", thinning, " iterations")
+
+    thinning <- seq(1, nrow(output), by=thinning)
+    output <- output[thinning,]
+    Us <- Us[thinning]
+    if(nrow(output) < min.its)
+      message(paste("Fewer than", min.its, "MCMC iterations remaining, please consider a longer run by increasing 'its'"))
+
+    rm(tw)
+    if(clean.garbage)
+      gc()
+
+    # save output and variables for future use
+    write.table(output, file.path(strat.dir, name, paste0(name, ".out")), row.names=FALSE, quote=FALSE, col.names=FALSE)
+    write.table(Us, file.path(strat.dir, name, paste0(name, "_energy.out")), row.names=FALSE, quote=FALSE, col.names=FALSE)
   }
-  message("Removed a burn-in of ", burnin)
 
-  # additional thinning after the run
-  if(length(thinning) == 0) # by default, take the thinning value as suggested by rtwalk's IAT
-    thinning <- round(max(1, coffee::IAT(tw, 0, burnin, nrow(output))))
-  message("Thinning the MCMC by storing every ", thinning, " iterations")
-
-  thinning <- seq(1, nrow(output), by=thinning)
-  output <- output[thinning,]
-  Us <- Us[thinning]
-  if(nrow(output) < min.its)
-    message(paste("Fewer than", min.its, "MCMC iterations remaining, please consider a longer run by increasing 'its'"))
-
-  rm(tw)
-  if(clean.garbage)
-    gc()
-
-  # save output and variables for future use
-  write.table(output, file.path(strat.dir, name, paste0(name, ".out")), row.names=FALSE, quote=FALSE, col.names=FALSE)
-  write.table(Us, file.path(strat.dir, name, paste0(name, "_energy.out")), row.names=FALSE, quote=FALSE, col.names=FALSE)
   info$dets <- dets
   info$dat <- dat
   info$struc <- struc
@@ -236,36 +293,45 @@ strat <- function(name="mystrat", strat.dir="strats", its=5e4, burnin=100, thinn
   assign_to_global("info", info)
 
   # draw the dates and relative information
-  if(length(dat[,5] == 10) > 0)
-    y.scale <- "positions"  
-  dates <- draw.strat(name, info, struc, BCAD=BCAD, strat.dir=strat.dir, y.scale=y.scale, cc.dir=cc.dir, postbomb=postbomb, ybottom.lab=y.scale, ...)
-
+  #if(length(dat[,5] == 10) > 0)
+  #  y.scale <- "positions"
+  dates <- draw.strat(name, info, struc, BCAD=BCAD, strat.dir=strat.dir, y.scale=y.scale, cc.dir=cc.dir, postbomb=postbomb, ybottom.lab=y.scale, min.its=min.its, ...)
+#  info <- draw.strat(name, info, struc, BCAD=BCAD, strat.dir=strat.dir, y.scale=y.scale, cc.dir=cc.dir, postbomb=postbomb, ybottom.lab=y.scale, ...)
+  info$dates <- dates
   if(length(struc$pos.gaps) > 0)
     for(i in rev(struc$pos.gaps))
       struc$pos.dates[struc$pos.dates>i] <- struc$pos.dates[struc$pos.dates>i] - 1 # why this?
-  overlapping <- overlap(dates$ages, dates$probs, output[,struc$pos.dates])
-  info$overlap <- overlapping
-  info$name <- name
+  within.hpds <- within.hpd(dates$ages, dates$probs, output[,struc$pos.dates], prob)
+  info$within.hpds <- within.hpds
   info$strat.dir <- file.path(strat.dir, name, name)
-  assign_to_global("info", info) # to ensure new information is available after running the strat function
+  
+  # to ensure new information is available after running the strat function
+  if(save.info)
+    assign_to_global("info", info) 
 
   if(talk) { 
-    o <- order(overlapping)
+    o <- order(within.hpds)
     pos.dates <- struc$pos.dates[o]
-    message("Mean overlap between calibrated and modelled dates ",
-      round(mean(overlapping),2), "%, ranging from ", 
-      round(min(overlapping),2), "% (date ", pos.dates[1], ") to ",
-      round(max(overlapping),2), "% (date ", pos.dates[length(pos.dates)], ")")
+    message(round(mean(within.hpds),2),
+	  "% of the model ages fit within the ", 100*prob, 
+	  "% hpd ranges of the dates, with worst-fitting date ", 
+	  pos.dates[1], " (",
+      round(min(within.hpds),2), 
+	  "%) and best-fitting date ", 
+	  pos.dates[length(pos.dates)], 
+	  " (", round(max(within.hpds),2), "%)")
 
     took <- as.numeric(format(Sys.time(), "%s")) - start.time
-    if(took < 60) 
-      message("Run took ", round(took, 2), " seconds") else
-      if(took < 3600)
-        message("Run took ", round(took/60, 2), " minutes") else
-        if(took < 86400)
-          message("Run took ", round(took/3600, 2), " hours") else
-            message("Run took ", round(took/86400, 2), " days")
-  } 
+    if(run)
+      if(took < 60)
+        message("Run took ", round(took, 2), " seconds") else
+        if(took < 3600)
+          message("Run took ", round(took/60, 2), " minutes") else
+          if(took < 86400)
+            message("Run took ", round(took/3600, 2), " hours") else
+              message("Run took ", round(took/86400, 2), " days")
+  }
+  invisible(info) 
 }
 
 
@@ -302,8 +368,8 @@ calib <- function(y, er, cc.y, cc.er, normal, ta, tb)
 # d(phi) = phi_0 - phi_m is the (modelled) span between the boundary ages phi_m and phi_0.
 # f(phi,theta) = ( 1/ (R - d(phi)) ) ( (1/ (d_phi)^(M-1) )
 # as log: -log((R - dphi) - ((M-1) * log(dphi)))
-l.span <- function(max.x, min.x, R, M) {
-  dphi <- abs(max.x - min.x)
+l.span <- function(x1, x2, R, M) {
+  dphi <- abs(x2 - x1)
   return(-log((R - dphi) - ((M-1) * log(dphi))))
 }
 
@@ -330,7 +396,25 @@ ages.undated <- function(position, set=get('info'), draw=TRUE) {
 
 
 
-# internal function to calculate how well the modelled distributions fit with the calibrated ones
+# for all model age estimates, check if they fall within one of the hpd ranges of the dates
+within.hpd <- function(calibs, probs, modelled, prob=.95) {
+  fits <- c()
+  # for each modelled age, find the height on the date's distribution (peak=1)
+  for(i in 1:ncol(calibs)) {
+    this.hpd <- rbind(hpd(cbind(calibs[,i], probs[,i]), prob))
+    inside <- rep(0, nrow(modelled))
+    for(j in 1:nrow(this.hpd)) { # check which model ages fit within any of the hpd ranges
+	  rng <- range(this.hpd[j,1:2])
+      inside <- inside + ((modelled[,i] > rng[1]) * (modelled[,i] < rng[2]))
+	}  
+	fits[i] <- length(which(inside > 0)) / nrow(modelled) # ratio
+  }
+  return(100*fits) # as percentage
+}
+
+
+
+# alternative internal function to calculate how well the modelled distributions fit with the calibrated ones
 overlap <- function(calibs, probs, modelled, n=1e3) {
   same <- c()
   for(i in 1:ncol(modelled)) {