+++ R/twalk.R
@@ -126,19 +126,19 @@ Runtwalk <- function(Tr, Obj, Supp, dat, dim = length(x0), x0=x0, xp0=xp0, at=6,
     Tr <- 0
   }
 
+  write.MCMC <- ifelse(length(out.fl) == 0, FALSE, TRUE) 
+
   every <- ceiling(Tr/thinning) # find how many sub-runs to run
+  if(write.MCMC)
+	message("writing results to temporary folder (see 'tempdir()')") else {
+      rec <- array(NA, dim=c(every, 2*c(1+length(x))))
+      recacc <- array(NA, dim=c(every, 2))
+    }  
+
   if(show.progress)
     pb <- txtProgressBar(min=0, max=Tr, style = 3, char=">")
 
-  if(length(out.fl) > 0) { # then we'll save the its in files while running
-    if(file.exists(out.fl))
-     file.remove(out.fl)
-    if(file.exists(energy.fl))
-      file.remove(energy.fl)
-  }
-	  
   acc <- 0
-  j <- 0
   for(i in 1:Tr) {
     if(show.progress)
       if(i %% 100)
@@ -155,23 +155,21 @@ Runtwalk <- function(Tr, Obj, Supp, dat, dim = length(x0), x0=x0, xp0=xp0, at=6,
          tmp.recacc <- c(move$funh, 0)
 
     # sub-runs to reduce the amount of iterations to be stored
-    j <- j+1
-    if(j == thinning) { # then store the iteration
-	  if(length(out.fl) > 0) {
-        fastwrite(t(c(U, Up, x, xp)), out.fl, append=TRUE, sep=",", row.names=FALSE, col.names=FALSE)		  
-        fastwrite(t(tmp.recacc), energy.fl, append=TRUE, sep=",", row.names=FALSE, col.names=FALSE)
-	  } else {
-        rec <- rbind(rec, c(U, Up, x, xp))
-        recacc <- rbind(recacc, tmp.recacc) # add
-      }
-      j <- 0 # reset
+    if(i %% thinning == 0) { # then store the iteration
+      if(write.MCMC) {  
+        fastwrite(t(c(U, Up, x, xp)), out.fl, append=TRUE, row.names=FALSE, col.names=FALSE)		  
+        fastwrite(t(tmp.recacc), energy.fl, append=TRUE, row.names=FALSE, col.names=FALSE)
+      } else {
+          rec[i/thinning,] <- c(U, Up, x, xp)
+          recacc[i/thinning,] <- tmp.recacc
+        }
     }
   }
   message("\nDone running...")
 
   if(length(out.fl) > 0) {
-    rec <- fastread(out.fl, header=FALSE, sep=",") 
-    recacc <- fastread(energy.fl, header=FALSE, sep=",")
+    rec <- fastread(out.fl, header=FALSE) 
+    recacc <- fastread(energy.fl, header=FALSE)
   }
 
   list( dim=dim, Tr=Tr, acc=acc, Us=rec[,1], Ups=rec[,2],
@@ -482,12 +480,12 @@ GetAutoCov <- function( dt, lags) {
 IAT <- function(set, par=0, from=1, to) {
   ## -lag/log(GetAutoCorr( info, lag, par=par, from=from, to=to))
   ## we get the desired time series, the parameter and the from - to selection
-  if(par>0) {
+  if(par > 0) {
     if(set$dim > 1)
       dt <- set$output[from:to, par] else
         dt <- set$output[from:to]
   } else
-    dt <-  set$Us[from:to]
+    dt <- set$Us[from:to]
 
   n <- to-from
   mu <- mean(dt)  ### with its mean and variance